Added working nutrient resource ui & Nutrient Resource System improvements
There is now a working visual for the nutrient reserves in the form of an auto adjusting "progress" type bar. If the bar drops to 0 it will not allow the player to create more roots, effectively being game over. Functionality was added to to the RootParts.cs and Player Controller to control nutrient reserve incrementing and decrementing